ElementClickInterceptedException Handled
I added a click_element button that adds modularity to my code, cleaning un the main method and allowing for reusability. It also handles the ElementClickInterceptedException by executing a java script command that scrolls to the top of the page and if that doesn't work, it scrolls to the element in question.

I also added a get_xpath method in the utils package which prints the entire XPATH of the element in question. This is especially useful with debugging and writing code.